+++ src/main/java/com/epam/esm/repository/order/OrderRepositoryImpl.java
@@ -1,6 +1,10 @@
 package com.epam.esm.repository.order;
 
+import com.epam.esm.entity.GiftCertificate;
 import com.epam.esm.entity.Order;
+import com.epam.esm.entity.Tag;
+import com.epam.esm.exception.EntityNotFoundException;
+import com.epam.esm.repository.CriteriaSpecification;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -19,6 +23,7 @@ import java.util.Optional;
 @Repository
 @RequiredArgsConstructor(onConstructor = @__(@Autowired))
 public class OrderRepositoryImpl implements OrderRepository {
+    private static final String ENTITY_NAME = "Order";
     @PersistenceContext
     private final EntityManager entityManager;
 
@@ -30,18 +35,24 @@ public class OrderRepositoryImpl implements OrderRepository {
     }
 
     @Override
-    public Optional<Order> remove(long orderId) {
-        Order order = entityManager.find(Order.class, orderId);
-        entityManager.remove(order);
-        entityManager.getTransaction().commit();
-        return Optional.of(order);
+    public Optional<Order> remove(long id) {
+        Order order = entityManager.find(Order.class, id);
+        if(order!=null) {
+            entityManager.remove(order);
+            entityManager.getTransaction().commit();
+            return Optional.of(order);
+        }
+        throw new EntityNotFoundException(ENTITY_NAME, id);
     }
 
     @Override
     public Optional<Order> findOrder(long id) {
         Order order = entityManager.find(Order.class, id);
-        entityManager.detach(order);
-        return Optional.of(order);
+        if (order != null) {
+            entityManager.detach(order);
+            return Optional.of(order);
+        }
+        throw new EntityNotFoundException(ENTITY_NAME, id);
     }
 
     @Override
@@ -54,5 +65,17 @@ public class OrderRepositoryImpl implements OrderRepository {
         allQuery.setFirstResult((page - 1) * pageSize);
         allQuery.setMaxResults(pageSize);
         return allQuery.getResultList();
+    }
+
+    @Override
+    public List<Order> findAllBySpecification(CriteriaSpecification<Order> specification, int page, int pageSize) {
+        CriteriaBuilder builder = entityManager.getCriteriaBuilder();
+        CriteriaQuery<Order> criteriaQuery = builder.createQuery(Order.class);
+        Root<Order> tagRoot = criteriaQuery.from(Order.class);
+        criteriaQuery.where(specification.toPredicate(tagRoot, builder));
+        TypedQuery<Order> query = entityManager.createQuery(criteriaQuery);
+        query.setFirstResult((page - 1) * pageSize);
+        query.setMaxResults(pageSize);
+        return query.getResultList();
     }
 }
